// main/front/types.ts
// Interfaces para tipar las respuestas de la API

export interface Author {
  name: string;
  lastname: string;
}

export interface Price {
  currency: string;
  amount: number;
  decimals: number;
}

export interface Product {
  id: string;
  title: string;
  price: Price;
  picture: string;
  condition: string;
  free_shipping: boolean;
}

export interface ProductDetail extends Product {
  sold_quantity: number;
  description: string;
}

export interface SearchResult {
  author: Author;
  categories: string[];
  items: Product[];
}

export interface Item {
  id: string;
  title: string;
  price: Price;
  picture: string;
  condition: string;
  free_shipping: boolean;
  city?: string;
}

export interface ItemDetails extends Item {
  sold_quantity: number;
  description: string;
}

export interface ProductDetailsResponse {
  author: Author;
  item: ProductDetail;
}